requirements.txt
fastapi==0.111.0
pydantic==2.8.1
pydantic-settings==2.3.4
starlette==0.37.2
uvicorn==0.30.1
email_validator==2.2.0
SQLAlchemy==2.0.31
asyncpg==0.29.0
PyJWT==2.8.0
passlib==1.7.4
